requirements.txt
-i https://pypi.org/simple
boto3==1.11.0
botocore==1.14.0
certifi==2019.11.28
chardet==3.0.4
click==7.0
colorlog==4.1.0
docutils==0.15.2
hydra-colorlog==0.1.4
hydra-core==0.11.3
idna==2.8
jmespath==0.9.4
joblib==0.14.1
numpy==1.18.1
omegaconf==1.4.1
python-dateutil==2.8.1
pyyaml==5.3
regex==2020.1.8
requests==2.22.0
s3transfer==0.3.0
sacremoses==0.0.38
sentencepiece==0.1.85
six==1.13.0
tokenizers==0.1.0
tqdm==4.41.1
transformers==2.3.0
urllib3==1.25.7
